Checking docker installation

+++ moteur-server-rest/workflow_manager.py
@@ -11,6 +11,13 @@ from config import get_workflow_filename
 
 logger = logging.getLogger(__name__)
 
+DOCKER_AVAILABLE = False
+
+def set_docker_available(status: bool):
+    global DOCKER_AVAILABLE
+    DOCKER_AVAILABLE = status
+    logger.info(f"Docker available: {DOCKER_AVAILABLE}")
+
 def find_process_pids(workflow_id):
     user = get_env_variable('USER', required=True)
     moteur_process_class = get_env_variable('MOTEUR_MAIN_CLASS', required=True)
@@ -99,18 +106,19 @@ def _kill_workflow(workflow_id, hard_kill):
         os.killpg(pgid, signal)
         logger.info(f"Process group {pgid} killed with signal {signal}.")
 
-        result = subprocess.run(
-            ["docker", "ps", "-q", "--filter", f"name={workflow_id}*"],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=False
-        )
-        if result.stdout.decode().strip():
-            container_ids = result.stdout.decode().strip().split('\n')
-            for container_id in container_ids:
-                if container_id.strip():
-                    os.system(f"docker kill {container_id.strip()}")
-                    logger.info(f"Container {container_id.strip()} killed.")
+        if DOCKER_AVAILABLE:
+            result = subprocess.run(
+                ["docker", "ps", "-q", "--filter", f"name={workflow_id}*"],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                check=False
+            )
+            if result.stdout.decode().strip():
+                container_ids = result.stdout.decode().strip().split('\n')
+                for container_id in container_ids:
+                    if container_id.strip():
+                        os.system(f"docker kill {container_id.strip()}")
+                        logger.info(f"Container {container_id.strip()} killed.")
     
     except subprocess.CalledProcessError as e:
         logger.warning(f"Error finding or killing process: {e}")
